Updated with non-forceful autocorrection update.
The previous command would always source the full entire list of iabbrev
statements. I wanted to put this in an autocommand for different
filetypes, but I didn't want to have it repeat this every time a new
file was opened. So the script now has a variable that checks whether
the list has been loaded.

I next want to update the names of the public functions to have more
consistency.